CCDM: re-format and re-locate test classes
