+++ flow-server/src/main/java/com/vaadin/flow/server/connect/VaadinServiceNameChecker.java
@@ -26,46 +26,52 @@ import java.util.regex.Pattern;
  * A checker responsible for validating the Vaadin Connect service names.
  */
 public class VaadinServiceNameChecker {
-  /**
-   * Set of reserved words in ECMAScript specification. Also covers all the
-   * reserved identifiers in TypeScript.
-   *
-   * @see <a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Lexical_grammar#Keywords">JavaScript Reference on MDN</a>
-   * @see <a href="https://github.com/Microsoft/TypeScript/blob/master/doc/TypeScript%20Language%20Specification.pdf">TypeScript Language Specification</a>
-   */
-  public static final Set<String> ECMA_SCRIPT_RESERVED_WORDS = Collections
-      .unmodifiableSet(new HashSet<>(Arrays.asList("abstract", "arguments",
-          "await", "boolean", "break", "byte", "case", "catch", "char", "class",
-          "const", "continue", "debugger", "default", "delete", "do", "double",
-          "else", "enum", "eval", "export", "extends", "false", "final",
-          "finally", "float", "for", "function", "goto", "if", "implements",
-          "import", "in", "instanceof", "int", "interface", "let", "long",
-          "native", "new", "null", "package", "private", "protected", "public",
-          "return", "short", "static", "super", "switch", "synchronized",
-          "this", "throw", "throws", "transient", "true", "try", "typeof",
-          "var", "void", "volatile", "while", "with", "yield")));
+    /**
+     * Set of reserved words in ECMAScript specification. Also covers all the
+     * reserved identifiers in TypeScript.
+     *
+     * @see <a href=
+     *      "https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Lexical_grammar#Keywords">JavaScript
+     *      Reference on MDN</a>
+     * @see <a href=
+     *      "https://github.com/Microsoft/TypeScript/blob/master/doc/TypeScript%20Language%20Specification.pdf">TypeScript
+     *      Language Specification</a>
+     */
+    public static final Set<String> ECMA_SCRIPT_RESERVED_WORDS = Collections
+            .unmodifiableSet(new HashSet<>(Arrays.asList("abstract",
+                    "arguments", "await", "boolean", "break", "byte", "case",
+                    "catch", "char", "class", "const", "continue", "debugger",
+                    "default", "delete", "do", "double", "else", "enum", "eval",
+                    "export", "extends", "false", "final", "finally", "float",
+                    "for", "function", "goto", "if", "implements", "import",
+                    "in", "instanceof", "int", "interface", "let", "long",
+                    "native", "new", "null", "package", "private", "protected",
+                    "public", "return", "short", "static", "super", "switch",
+                    "synchronized", "this", "throw", "throws", "transient",
+                    "true", "try", "typeof", "var", "void", "volatile", "while",
+                    "with", "yield")));
 
-  private static final Pattern CONTAINS_WHITESPACE_PATTERN = Pattern
-      .compile(".*[\\s+].*");
+    private static final Pattern CONTAINS_WHITESPACE_PATTERN = Pattern
+            .compile(".*[\\s+].*");
 
-  /**
-   * Validates the Vaadin Connect service name given.
-   *
-   * @param serviceName
-   *          the name to validate
-   * @return {@code null} if there are no validation errors or {@link String}
-   *         containing validation error description.
-   */
-  public String check(String serviceName) {
-    if (serviceName == null || serviceName.isEmpty()) {
-      return "Service name cannot be blank";
+    /**
+     * Validates the Vaadin Connect service name given.
+     *
+     * @param serviceName
+     *            the name to validate
+     * @return {@code null} if there are no validation errors or {@link String}
+     *         containing validation error description.
+     */
+    public String check(String serviceName) {
+        if (serviceName == null || serviceName.isEmpty()) {
+            return "Service name cannot be blank";
+        }
+        if (ECMA_SCRIPT_RESERVED_WORDS.contains(serviceName)) {
+            return "Service name cannot be equal to JavaScript reserved words";
+        }
+        if (CONTAINS_WHITESPACE_PATTERN.matcher(serviceName).matches()) {
+            return "Service name cannot contain any whitespaces";
+        }
+        return null;
     }
-    if (ECMA_SCRIPT_RESERVED_WORDS.contains(serviceName)) {
-      return "Service name cannot be equal to JavaScript reserved words";
-    }
-    if (CONTAINS_WHITESPACE_PATTERN.matcher(serviceName).matches()) {
-      return "Service name cannot contain any whitespaces";
-    }
-    return null;
-  }
 }
