+++ flow-server/src/test/java/com/vaadin/flow/server/connect/VaadinServiceNameCheckerTest.java
@@ -27,53 +27,53 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 public class VaadinServiceNameCheckerTest {
-  private final VaadinServiceNameChecker checker = new VaadinServiceNameChecker();
+    private final VaadinServiceNameChecker checker = new VaadinServiceNameChecker();
 
-  @Test
-  public void should_ReturnValidationError_When_NullAsServiceNameGiven() {
-    String validationResult = checker.check(null);
+    @Test
+    public void should_ReturnValidationError_When_NullAsServiceNameGiven() {
+        String validationResult = checker.check(null);
 
-    assertNotNull(validationResult);
-    assertTrue(validationResult.contains("blank"));
-  }
-
-  @Test
-  public void should_ReturnValidationError_When_EmptyAsServiceNameGiven() {
-    String validationResult = checker.check("");
-
-    assertNotNull(validationResult);
-    assertTrue(validationResult.contains("blank"));
-  }
-
-  @Test
-  public void should_ReturnValidationError_When_ReservedWordAsServiceNameGiven() {
-    for (String reservedEsWord : VaadinServiceNameChecker.ECMA_SCRIPT_RESERVED_WORDS) {
-      String validationResult = checker.check(reservedEsWord);
-
-      assertNotNull(String.format(
-          "Word '%s' should cause validation errors, but it did not",
-          reservedEsWord), validationResult);
-      assertTrue(validationResult.contains("reserved"));
+        assertNotNull(validationResult);
+        assertTrue(validationResult.contains("blank"));
     }
-  }
 
-  @Test
-  public void should_ReturnValidationError_When_ServiceNameWithWhitespaceGiven() {
-    for (String wordWithWhitespace : Arrays.asList("bad name", "bad\nname",
-        "bad\tname", "bad\rname")) {
-      String validationResult = checker.check(wordWithWhitespace);
+    @Test
+    public void should_ReturnValidationError_When_EmptyAsServiceNameGiven() {
+        String validationResult = checker.check("");
 
-      assertNotNull(String.format(
-          "Word '%s' should cause validation errors, but it did not",
-          wordWithWhitespace), validationResult);
-      assertTrue(validationResult.contains("whitespace"));
+        assertNotNull(validationResult);
+        assertTrue(validationResult.contains("blank"));
     }
-  }
 
-  @Test
-  public void should_ReturnNoErrors_When_ValidServiceNameGiven() {
-    String validationResult = checker.check("ordinaryName");
+    @Test
+    public void should_ReturnValidationError_When_ReservedWordAsServiceNameGiven() {
+        for (String reservedEsWord : VaadinServiceNameChecker.ECMA_SCRIPT_RESERVED_WORDS) {
+            String validationResult = checker.check(reservedEsWord);
 
-    assertNull(validationResult);
-  }
+            assertNotNull(String.format(
+                    "Word '%s' should cause validation errors, but it did not",
+                    reservedEsWord), validationResult);
+            assertTrue(validationResult.contains("reserved"));
+        }
+    }
+
+    @Test
+    public void should_ReturnValidationError_When_ServiceNameWithWhitespaceGiven() {
+        for (String wordWithWhitespace : Arrays.asList("bad name", "bad\nname",
+                "bad\tname", "bad\rname")) {
+            String validationResult = checker.check(wordWithWhitespace);
+
+            assertNotNull(String.format(
+                    "Word '%s' should cause validation errors, but it did not",
+                    wordWithWhitespace), validationResult);
+            assertTrue(validationResult.contains("whitespace"));
+        }
+    }
+
+    @Test
+    public void should_ReturnNoErrors_When_ValidServiceNameGiven() {
+        String validationResult = checker.check("ordinaryName");
+
+        assertNull(validationResult);
+    }
 }
