+++ flow-server/src/test/java/com/vaadin/flow/server/connect/typeconversion/BeanConversionIT.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2000-2019 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.vaadin.flow.server.connect.typeconversion;
+
+import org.junit.Test;
+
+public class BeanConversionIT extends BaseTypeConversionIT {
+
+    @Test
+    public void should_ConvertToBean_When_ReceiveBeanObject() {
+        String inputValue = "{\"name\":\"mybean\",\"address\":\"myaddress\","
+                + "\"age\":10,\"isAdmin\":true,\"testEnum\":\"FIRST\","
+                + "\"roles\":[\"Admin\"], \"customProperty\": \"customValue\"}";
+        String expectedValue = "{\"name\":\"mybean-foo\","
+                + "\"address\":\"myaddress-foo\","
+                + "\"age\":11,\"isAdmin\":false," + "\"testEnum\":\"SECOND\","
+                + "\"roles\":[\"Admin\",\"User\"],\"customProperty\":\"customValue-foo\"}";
+        assertEqualExpectedValueWhenCallingMethod("getFooBean", inputValue,
+                expectedValue);
+    }
+}
